requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
django-cors-headers==3.6.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
futures==3.1.1
goslate==1.5.1
idna==2.10
joblib==1.0.1
nltk==3.5
PyDictionary==2.0.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
tqdm==4.58.0
urllib3==1.26.3
numpy~=1.20.2
opencv-python~=4.5.1.48
Pillow~=8.2.0
